reviewer/requirements.txt
flask>=2.0.0
google-cloud-firestore>=2.4.0
google-cloud-storage>=2.0.0